Move ScrollOffsetNotifier to own file
