+++ packages/scrollable_positioned_list/lib/src/scroll_offset_listener.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
 
+import 'scroll_offset_notifier.dart';
+
 /// Provides an affordance for listening to scroll offset changes.
 abstract class ScrollOffsetListener {
   /// Stream of scroll offset deltas.
@@ -15,19 +17,3 @@ abstract class ScrollOffsetListener {
           recordProgrammaticScrolls: recordProgrammaticScrolls);
 }
 
-class ScrollOffsetNotifier implements ScrollOffsetListener {
-  final bool recordProgrammaticScrolls;
-
-  ScrollOffsetNotifier({this.recordProgrammaticScrolls = true});
-
-  final _streamController = StreamController<double>();
-
-  @override
-  Stream<double> get changes => _streamController.stream;
-
-  StreamController get changeController => _streamController;
-
-  void dispose() {
-    _streamController.close();
-  }
-}
